# openapi.yaml
openapi: 3.1.0
info:
  title: Samokoder SaaS API
  version: 1.0.0
  description: |
    Полнофункциональный API для платформы Samokoder - AI-powered генератор кода.
    
    ## Аутентификация
    API использует JWT токены с Bearer схемой аутентификации. Токены передаются в заголовке:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Безопасность
    - JWT токены с ротацией (access: 15 минут, refresh: 7 дней)
    - Rate limiting на критических эндпоинтах
    - Защита от brute-force атак
    - CORS с белым списком доменов
    - Валидация всех входных данных
    
    ## Версионирование
    API версионируется через префикс `/v1`. Мы гарантируем обратную совместимость в рамках мажорной версии.
    
  contact:
    name: Samokoder Support
    email: support@samokoder.io
  license:
    name: Proprietary
servers:
  - url: https://api.samokoder.io
    description: Production сервер
  - url: https://staging-api.samokoder.io
    description: Staging сервер
  - url: http://localhost:8000
    description: Локальная разработка
...
paths:
...
  /v1/analytics/user:
    get:
      tags:
        - analytics
      summary: Получить аналитику пользователя
      description: |
        Возвращает метрики аналитики для текущего пользователя.
        
        **Реализация:** `api/routers/analytics.py:12-31`
      operationId: getUserAnalytics
      responses:
        '200':
          description: Метрики пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                  metrics:
                    type: object
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Ошибка получения аналитики
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /v1/analytics/system:
    get:
      tags:
        - analytics
      summary: Получить системную аналитику
      description: |
        Возвращает общую системную аналитику (требует права администратора).
        
        **Реализация:** `api/routers/analytics.py:34-56`
        
        **⚠️ TODO:** Проверка прав администратора не реализована
      operationId: getSystemAnalytics
      responses:
        '200':
          description: Системные метрики
          content:
            application/json:
              schema:
                type: object
                properties:
                  metrics:
                    type: object
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Ошибка получения аналитики
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /v1/analytics/user/actions:
    get:
      tags:
        - analytics
      summary: Получить лог действий пользователя
      description: |
        Возвращает лог действий текущего пользователя.
        
        **Реализация:** `api/routers/analytics.py:59-80`
      operationId: getUserActionLogs
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
          description: Максимальное количество записей
      responses:
        '200':
          description: Лог действий
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                  logs:
                    type: array
                    items:
                      type: object
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Ошибка получения логов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /v1/analytics/record:
    post:
      tags:
        - analytics
      summary: Записать действие пользователя
      description: |
        Записывает действие пользователя для аналитики.
        
        **Реализация:** `api/routers/analytics.py:83-108`
      operationId: recordUserAction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                details:
                  type: object
                project_id:
                  type: string
              required:
                - action
      responses:
        '200':
          description: Действие записано
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Ошибка записи
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
  /v1/analytics/export:
    get:
      tags:
        - analytics
      summary: Экспортировать аналитику
      description: |
        Экспортирует всю аналитику (требует права администратора).
        
        **Реализация:** `api/routers/analytics.py:111-131`
        
        **⚠️ TODO:** Проверка прав администратора не реализована
      operationId: exportAnalytics
      responses:
        '200':
          description: Экспортированные данные
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Ошибка экспорта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT токен в формате: Bearer <token>
        
        Получить токен можно через `/v1/auth/login` или `/v1/auth/register`
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Описание ошибки
        error_code:
          type: string
          description: Код ошибки (опционально)
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
...
security:
  - BearerAuth: []
...